feat: add layers controller
